fix(build): improve cross-platform compatibility for socket reuse and signal tests
