+++ cmd/listen_other.go
@@ -1,0 +1,9 @@
+//go:build !unix && !linux && !darwin && !dragonfly && !freebsd && !netbsd && !openbsd && !solaris && !aix
+
+package cmd
+
+import "syscall"
+
+func setReusePort(c syscall.RawConn) error {
+	return nil
+}
